# rest/swagger/swagger.json
{"swagger":"2.0","info":{"title":"Guest Book API","description":"API for Guest Book","version":"1.0"},"schemes":["https"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/guestbook/":{"get":{"tags":["guestbook"],"summary":"List of guests","operationId":"guestbook#list","produces":["application/vnd.guestbook.error+json","application/vnd.guestbook.guest+json; type=collection"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GuestbookGuestCollection"}},"400":{"description":"BadRequest is returned if input object is missing required attributes or their values are out of range.","schema":{"$ref":"#/definitions/GuestbookError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/GuestbookError"}}},"schemes":["https"]},"post":{"tags":["guestbook"],"summary":"Add an entry","operationId":"guestbook#add","produces":["application/vnd.guestbook.error+json"],"parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/guestbookType"}}],"responses":{"204":{"description":"No Content"},"400":{"description":"BadRequest is returned if input object is missing required attributes or their values are out of range.","schema":{"$ref":"#/definitions/GuestbookError"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/GuestbookError"}}},"schemes":["https"]}},"/guestbook/swagger.json":{"get":{"summary":"Download rest/swagger/swagger.json","operationId":"swagger#/guestbook/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["https"]}},"/guestbook/version":{"get":{"tags":["version"],"summary":"Return application's version and commit hash","operationId":"version#version","produces":["application/vnd.guestbook.version+json"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GuestbookVersion"}}},"schemes":["https"]}},"/guestbook/{id}":{"delete":{"tags":["guestbook"],"summary":"Delete an entry","operationId":"guestbook#delete","produces":["application/vnd.guestbook.error+json"],"parameters":[{"name":"id","in":"path","description":"Entry ID","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"},"400":{"description":"BadRequest is returned if input object is missing required attributes or their values are out of range.","schema":{"$ref":"#/definitions/GuestbookError"}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/GuestbookError"}}},"schemes":["https"]}}},"definitions":{"GuestbookError":{"title":"Mediatype identifier: application/vnd.guestbook.error+json; view=default","type":"object","properties":{"code":{"type":"string","description":"Message ID","example":"UNAUTHORIZED"},"msg":{"type":"string","description":"Localized message","example":"Unauthorized access"}},"description":"GuestbookError media type (default view)","example":{"code":"UNAUTHORIZED","msg":"Unauthorized access"},"required":["code","msg"]},"GuestbookGuest":{"title":"Mediatype identifier: application/vnd.guestbook.guest+json; view=default","type":"object","properties":{"comment":{"type":"string","description":"Comment","example":"Lorem ipsum"},"id":{"type":"string","description":"Guestbook ID","example":"c2d9ecce-f6a3-49cf-8c71-7d4415beb3b8"},"timestamp":{"type":"number","description":"Entry timestamp","example":1543981510,"format":"double"}},"description":"GuestbookGuest media type (default view)","example":{"comment":"Lorem ipsum","id":"c2d9ecce-f6a3-49cf-8c71-7d4415beb3b8","timestamp":1543981510},"required":["id","timestamp","comment"]},"GuestbookGuestCollection":{"title":"Mediatype identifier: application/vnd.guestbook.guest+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/GuestbookGuest"},"description":"GuestbookGuestCollection is the media type for an array of GuestbookGuest (default view)","example":[{"comment":"Lorem ipsum","id":"c2d9ecce-f6a3-49cf-8c71-7d4415beb3b8","timestamp":1543981510},{"comment":"Lorem ipsum","id":"c2d9ecce-f6a3-49cf-8c71-7d4415beb3b8","timestamp":1543981510},{"comment":"Lorem ipsum","id":"c2d9ecce-f6a3-49cf-8c71-7d4415beb3b8","timestamp":1543981510}]},"GuestbookVersion":{"title":"Mediatype identifier: application/vnd.guestbook.version+json; view=default","type":"object","properties":{"git":{"type":"string","description":"Git commit hash","example":"000000"},"version":{"type":"string","description":"Application version","example":"1.0"}},"description":"GuestbookVersion media type (default view)","example":{"git":"000000","version":"1.0"},"required":["version"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}},"guestbookType":{"title":"guestbookType","type":"object","properties":{"comment":{"type":"string","description":"Comment","example":"Lorem ipsum"}},"example":{"comment":"Lorem ipsum"},"required":["comment"]}},"responses":{"BadRequest":{"description":"BadRequest is returned if input object is missing required attributes or their values are out of range.","schema":{"$ref":"#/definitions/error"}},"Forbidden":{"description":"Forbidden is returned when user is not authorized to perform an action.","schema":{"$ref":"#/definitions/error"}},"NoContent":{"description":"No Content"},"NotFound":{"description":"Not Found"},"Unauthorized":{"description":"Unauthorized is returned when user request does not contain authentication token or authentication is invalid. The response must include a valid \"WWW-Authenticate\" header.","schema":{"$ref":"#/definitions/error"},"headers":{"WWW-Authenticate":{"description":"https://tools.ietf.org/html/rfc7235","type":"string","default":"Bearer"}}}}}